# docs/api/openapi.bundle.yaml
openapi: 3.0.0
info:
  title: Lazuardy API Documentation
  description: Dokumentasi API untuk backend Lazuardy
  version: 1.0.0
  license:
    name: Proprietary
    url: http://127.0.0.1:8000
servers:
  - url: http://127.0.0.1:8000/api
    description: Local Server
...
paths:
  /registerOtpEmail:
    post:
      tags:
        - Authentication
      security: []
      summary: Mengirim OTP untuk Registrasi
      operationId: registerOtpEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
      responses:
        '201':
          description: OTP Berhasil dikirim
          content:
            application/json:
              example:
                status: success
        '422':
          description: Validasi gagal saat mengirim OTP registrasi
          content:
            application/json:
              example:
                message: The email field is required.
                errors:
                  email:
                    - The email field is required.
        '500':
          description: Gagal mengirim OTP. Status HTTP mengikuti exception code, dengan fallback 500.
          content:
            application/json:
              example:
                status: error
                message: Gagal mengirim OTP
                debug_error: SMTP Error...
  /verifyOtpRegisterEmail:
    post:
      tags:
        - Authentication
      security: []
      summary: Verifikasi OTP untuk Registrasi
      operationId: verifyOtpRegisterEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - otp
              properties:
                email:
                  type: string
                  format: email
                otp:
                  type: string
                  example: '1234'
      responses:
        '200':
          description: Verifikasi berhasil/selesai
          content:
            application/json:
              example:
                status: success
        '403':
          description: OTP tidak valid
          content:
            application/json:
              example:
                status: error
                message: Gagal memverifikasi OTP
                debug_error: Kode OTP yang Anda masukkan salah.
        '404':
          description: OTP tidak ditemukan atau kadaluwarsa
          content:
            application/json:
              example:
                status: error
                message: Gagal memverifikasi OTP
                debug_error: OTP tidak ditemukan atau sudah kedaluwarsa.
        '422':
          $ref: '#/components/responses/ValidationError'
  /forgotPasswordOtpEmail:
    post:
      tags:
        - Authentication
      security: []
      summary: Mengirim OTP untuk Lupa Password
      operationId: forgotPasswordOtpEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '201':
          description: OTP Berhasil dikirim
          content:
            application/json:
              example:
                status: success
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          description: Gagal mengirim OTP
          content:
            application/json:
              example:
                status: error
                message: Gagal mengirim OTP
                debug_error: Terjadi kesalahan sistem saat memproses OTP.
  /forgotPasswordVerifyOtpEmail:
    post:
      tags:
        - Authentication
      security: []
      summary: Verifikasi OTP untuk Lupa Password
      operationId: forgotPasswordVerifyOtpEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - otp
              properties:
                email:
                  type: string
                  format: email
                otp:
                  type: string
      responses:
        '200':
          description: Verifikasi berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  reset_token:
                    type: string
                    example: PmBbJbJTiub0IbNe
        '403':
          description: OTP salah
          content:
            application/json:
              example:
                status: error
                message: Gagal memverifikasi OTP
                debug_error: Kode OTP yang Anda masukkan salah.
        '404':
          description: OTP tidak ditemukan atau sudah kedaluwarsa
          content:
            application/json:
              example:
                status: error
                message: Gagal memverifikasi OTP
                debug_error: OTP tidak ditemukan atau sudah kedaluwarsa.
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          description: Internal Server Error
          content:
            application/json:
              example:
                status: error
                message: Gagal memverifikasi OTP
                debug_error: Terjadi kesalahan sistem saat memproses OTP.
  /forgotPasswordResetPassword:
    post:
      tags:
        - Authentication
      security: []
      summary: Reset password menggunakan token hasil verifikasi OTP
      operationId: forgotPasswordResetPassword
      description: |
        Endpoint ini digunakan untuk memperbarui password user setelah berhasil memverifikasi OTP.
        Token yang digunakan harus sesuai dengan yang disimpan di Cache pada tahap verifikasi.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - reset_token
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  description: Password baru minimal 8 karakter
                  minLength: 8
                  example: passwordBaru123
                reset_token:
                  type: string
                  description: Token unik yang didapat dari response forgotPasswordVerifyOtpEmail
                  example: aB1cD2eF3gH4iJ5k
      responses:
        '200':
          description: Password berhasil diperbarui
          content:
            application/json:
              example:
                status: success
        '400':
          description: Bad Request - Kesalahan umum lainnya
          content:
            application/json:
              example:
                status: error
                debug_error: User tidak ditemukan.
        '403':
          description: Forbidden - Token tidak valid atau sudah kedaluwarsa
          content:
            application/json:
              example:
                status: error
                debug_error: Token reset tidak valid atau sudah kedaluwarsa.
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /studentRegister:
    post:
      tags:
        - Authentication
      security: []
      summary: Mendaftarkan user siswa baru
      operationId: studentRegister
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentRegisterRequest'
      responses:
        '200':
          description: Berhasil mendaftar
          content:
            application/json:
              example:
                status: success
                access_token: 1|...
                token_type: Bearer
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          description: Gagal mendaftar
          content:
            application/json:
              example:
                status: error
                message: Gagal melakukan registrasi siswa.
                debug_error: Terjadi kesalahan saat menyimpan data ke sistem.
  /tutorRegister:
    post:
      tags:
        - Authentication
      security: []
      summary: Mendaftarkan user tutor baru (dengan upload file)
      operationId: tutorRegister
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - email
                - password
                - name
                - gender
                - description
                - date_of_birth
                - telephone_number
                - bank_code
                - account_number
                - province
                - regency
                - district
                - subdistrict
                - subject_id
                - learning_method
                - schedules
                - curriculum_vitae
                - id_card
                - certificate
              properties:
                email:
                  type: string
                  format: email
                  example: tutor@example.com
                password:
                  type: string
                  format: password
                  example: password123
                name:
                  type: string
                  example: Lazuardy Tutor
                gender:
                  type: string
                  enum:
                    - male
                    - female
                  example: male
                description:
                  type: string
                  example: Saya adalah guru matematika berpengalaman.
                date_of_birth:
                  type: string
                  format: date
                  example: '1995-01-01'
                telephone_number:
                  type: string
                  example: '08123456789'
                bank_code:
                  type: string
                  example: ID_BSI
                account_number:
                  type: string
                  example: '7123456789'
                profile_photo:
                  type: string
                  format: binary
                  description: Upload foto profil (jpg/png/jpeg/svg/webp)
                province:
                  type: string
                  example: Jawa Timur
                regency:
                  type: string
                  example: Malang
                district:
                  type: string
                  example: Lowokwaru
                subdistrict:
                  type: string
                  example: Tulusrejo
                google_id:
                  type: string
                  nullable: true
                facebook_id:
                  type: string
                  nullable: true
                subject_id:
                  type: integer
                  example: 1
                learning_method:
                  type: array
                  items:
                    type: string
                  example:
                    - online
                    - offline
                schedules:
                  type: array
                  description: Kirim sebagai array of objects
                  items:
                    type: object
                    properties:
                      day:
                        type: string
                        example: Monday
                      time:
                        type: string
                        example: '14:00'
                curriculum_vitae:
                  type: string
                  format: binary
                  description: Upload file PDF
                id_card:
                  type: string
                  format: binary
                  description: Upload file PDF/JPG/PNG
                certificate:
                  type: string
                  format: binary
                  description: Upload file PDF
      responses:
        '200':
          description: Berhasil mendaftar
          content:
            application/json:
              example:
                status: success
                access_token: 1|...
                token_type: Bearer
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          description: Gagal melakukan registrasi tentor.
          content:
            application/json:
              example:
                status: error
                message: Gagal melakukan registrasi tentor.
                debug_error: Terjadi kesalahan saat menyimpan data ke sistem.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ValidationError:
      description: Validasi Gagal
      content:
        application/json:
          example:
            message: Data tidak valid
            errors:
              email:
                - Email sudah terdaftar
...
  schemas:
...
    StudentRegisterRequest:
      type: object
      required:
        - email
        - password
        - name
        - class_id
        - gender
        - date_of_birth
        - telephone_number
        - province
        - regency
        - district
        - subdistrict
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        name:
          type: string
        class_id:
          type: integer
        gender:
          type: string
          enum:
            - male
            - female
        date_of_birth:
          type: string
          format: date
        telephone_number:
          type: string
        province:
          type: string
        regency:
          type: string
        district:
          type: string
        subdistrict:
          type: string
        google_id:
          type: string
          nullable: true
          example: null
        facebook_id:
          type: string
          nullable: true
          example: null
...